+++ app/Providers/AppServiceProvider.php
@@ -2,6 +2,7 @@
 
 namespace App\Providers;
 
+use Illuminate\Pagination\Paginator;
 use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Facades\Schema;
 
@@ -24,6 +25,9 @@ class AppServiceProvider extends ServiceProvider
      */
     public function boot()
     {
+        // Added for result pagination css. This uses Bootstrap css instead of the default Tailwind css.
+        Paginator::useBootstrap();
+
         // Added for the sake of old MariaDB version. Link to issue found below.
         // https://stackoverflow.com/questions/42244541/laravel-migration-error-syntax-error-or-access-violation-1071-specified-key-wa 
         Schema::defaultStringLength(191);