+++ resources/js/components/Bible/SearchResultsComponent.vue
@@ -1,15 +1,15 @@
 <template>
-    <div class="container">
+    <div class="container-fluid">
         <div class="row justify-content-center">
-            <div class="col-md-8">
+            <div class="col-12">
                 <div class="card">
                     <div class="card-header">Search Results</div>
 
                     <div class="card-body">
-                        {{ results }} search results found.
-                        <article v-for="result in results" :key="result.id">
-                            <!-- <p>{{ result.chapter.book.book }} {{ result.chapter.book_chapter }}:{{ result.chapter_verse }} - <span :v-html="result.verse"></span></p> -->
-                            <!-- <p>{{ result.chapter.book_chapter }}:{{ result.chapter_verse }} - <span :v-html="result.verse"></span></p> -->
+                        <article v-for="result in searchResults" :key="result.id">
+                            <p><a :href="'/' + result.book.toString() + '/' + result.book_chapter.toString() + '#' + 'verse-' + result.chapter_verse.toString()">
+                                {{ result.book }} {{ result.book_chapter }}:{{ result.chapter_verse }}
+                            </a>: - <span v-html="result.verse"></span></p>
                         </article>
                     </div>
                 </div>
@@ -20,9 +20,46 @@
 
 <script>
     export default {
-        mounted() {
-            console.log(this.results);
+        computed: {
+            searchResults: function() {
+                let tmpResults = [];
+                for(let i = 0; i < this.tokens.length; i++) {
+                    let searchContextRegEx = new RegExp(`${this.tokens[i]}`, 'ig');
+    
+                    for(let j = 0; j < this.results.length; j++) {
+                        tmpResults[j] = this.results[j];
+                        this.results[j].verse = this.results[j].verse.replaceAll(searchContextRegEx, `<span class='search-context'>${this.tokens[i]}</span>`);
+                    }
+                }
+                return tmpResults;
+            }
         },
-        props: ['results']
+        created: function() {
+        //     this.tokens.forEach(function(element) {
+        //         let searchContextRegEx = new RegExp(`/${element}/`);
+        //         this.results.forEach(function(verse) {
+        //             verse.replaceAll(searchContextRegEx, `<span class='search-context'>${element}</span>`);
+        //         });
+        //     });
+        // },
+        // mounted() {
+            console.log(this.tokens);
+            // this.tokens.forEach(function(element) {
+            //     let searchContextRegEx = new RegExp(`/${element}/`);
+            //     results.forEach(function(verse) {
+            //         verse.replaceAll(searchContextRegEx, `<span class='search-context'>${element}</span>`);
+            //     });
+            // });
+            for(let i = 0; i < this.tokens.length; i++) {
+                let searchContextRegEx = new RegExp(`${this.tokens[i]}`, 'ig');
+                // this.results.forEach(element => {
+                //     element.verse.replaceAll(searchContextRegEx, `<span class='search-context'>${this.tokens[i]}</span>`);
+                // });
+                for(let j = 0; j < this.results.length; j++) {
+                    this.results[j].verse = this.results[j].verse.replaceAll(searchContextRegEx, `<span class='search-context'>${this.tokens[i]}</span>`);
+                }
+            }
+        },
+        props: ['tokens', 'results']
     }
 </script>